requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.5.0
astroid==3.2.4
Authlib==1.3.1
bandit==1.7.9
base58==2.1.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
black==24.8.0
blis==0.7.11
boto3==1.35.0
boto3-stubs==1.35.0
botocore==1.35.0
botocore-stubs==1.35.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
commitizen==3.29.0
confection==0.1.5
cymem==2.0.8
decli==0.6.2
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dparse==0.6.4b0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.12.4
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
identify==2.6.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
jiter==0.5.0
jmespath==1.0.1
langcodes==3.4.0
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mccabe==0.7.0
mdurl==0.1.2
murmurhash==1.0.10
mypy==1.11.1
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
openai==1.41.0
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pandas-stubs==2.2.2.240807
paramiko==3.4.1
pathspec==0.12.1
pbr==6.0.0
platformdirs==4.2.2
plotly==5.24.0
pluggy==1.5.0
pre-commit==3.8.0
preshed==3.0.9
prompt-toolkit==3.0.36
psutil==6.0.0
pycoingecko==3.1.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pylint==3.2.6
PyNaCl==1.5.0
PyPika==0.48.9
pytest==8.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytoolconfig==1.3.1
pytz==2024.1
PyYAML==6.0.2
questionary==2.0.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.6.1
s3transfer==0.10.2
safety==3.2.5
safety-schemas==0.0.3
schedule==1.2.2
shellingham==1.5.4
six==1.16.0
selenium==4.25.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
srsly==2.4.8
stevedore==5.2.0
tenacity==9.0.0
termcolor==2.4.0
thefuzz==0.22.1
thinc==8.2.5
tiktoken==0.7.0
tldextract==5.1.2
tomlkit==0.13.2
tqdm==4.66.5
typer==0.12.4
types-awscrt==0.21.2
types-docutils==0.21.0.20240724
types-Pygments==2.18.0.20240506
types-pytz==2024.1.0.20240417
types-requests==2.32.0.20240712
types-s3transfer==0.10.1
types-setuptools==71.1.0.20240818
types-tqdm==4.66.0.20240417
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webdriver-manager==4.0.2
wrapt==1.16.0